server/sync/requirements.txt
pandas>=2.0
requests>=2.31
supabase>=2.0
python-dotenv>=1.0
